pyproject.toml
[tool.poetry]
name = "recommend-service"
version = "0.1.0"
description = ""
authors = [
    "Anton Shtarev <a.o.shtarev@gmail.com>",
    ]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.4"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.3.3"
grpcio = "^1.63.0"
grpcio-reflection = "^1.63.0"
torch = "^2.6.0"
onnxruntime = "^1.20.1"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
black = "^24.4.2"
pytest = "^8.2.2"
pytest-mock = "^3.14.0"
pyright = "^1.1.388"
onnx = "^1.17.0"


[tool.poetry.group.dev]
optional = true

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
src = ["src"]
line-length = 120
target-version = "py311"
# Group violations by containing file.
output-format = "grouped"
# Exclude
exclude = [
    "**/migrations/*.py",
    "*pb2*.py*",
]

[tool.ruff.lint]
# Disable warinings on this tags.
task-tags = ["TODO"]
# Rules.
select = [
    # pycodestyle.
    "E", "W",
    # Pyflakes.
    "F",
    # isort.
    "I",
    # pydocstyle.
    "D",
    # pyupgrade.
    "UP",
    # pep8-naming.
    "N",
    # flake8-2020.
    "YTT",
    # flake8-annotations.
    "ANN",
    # flake8-bandit.
    "S",
    # flake8-blind-except.
    "BLE",
    # flake8-boolean-trap.
    "FBT",
    # flake8-bugbear.
    "B",
    # flake8-builtins.
    "A",
    # flake8-comprehensions.
    "C4",
    # flake8-debugger.
    "T10",
    # flake8-errmsg.
    "EM",
    # flake8-implicit-str-concat.
    "ISC",
    # flake8-import-conventions.
    "ICN",
    # flake8-print.
    "T20",
    # flake8-pytest-style.
    "PT",
    # flake8-quotes.
    "Q",
    # flake8-return.
    "RET",
    # flake8-simplify.
    "SIM",
    # flake8-tidy-imports.
    "TID",
    # flake8-datetimez.
    "DTZ",
    # eradicate.
    "ERA",
    # pandas-vet.
    "PD",
    # pygrep-hooks.
    "PGH",
    # Pylint.
    "PLC", "PLE", "PLR", "PLW",
    # flake8-pie.
    "PIE",
    # Ruff-specific rules.
    "RUF",
]
# Disable.
ignore = [
    # Missing docstring in public module.
    "D100",
    # Missing docstring in public class.
    "D101",
    # Missing docstring in public method.
    "D102",
    # Missing docstring in public function.
    "D103",
    # Missing docstring in public package.
    "D104",
    # Missing docstring in public magic method.
    "D105",
    # Missing docstring in public nested class.
    "D106",
    # Missing docstring in __init__ method.
    "D107",
    # One-line docstring should fit on one line.
    "D200",
    # Multi-line docstring summary should start at the first line,
    "D212",
    # Missing type annotation for self in method.
    "ANN101",
    # Missing type annotation for cls in classmethod.
    "ANN102",
    # Dynamically typed expressions (typing.Any) are disallowed.
    "ANN401",
    # Shorted imports (numpy as np, etc).
    "ICN001",
    # Import from {module} instead: {names}
    "UP035",
    # Do not catch blind exception: `Exception`.
    "BLE001",
    # Possible SQL injection vector through string-based query construction.
    "S608",
    # Empty string in if.
    "PLC1901",
    # Weired safety issues
    "S311",
    "S603",
    "S602",
    "S301",
    "S104",
    # Class attribure `id`
    "A003",
    # Format for function calls
    "D415",
    "RUF012",
    # Allow assert for type checking
    "S101",
]

[tool.ruff.lint.isort]
# Combines "as" imports on the same line.
combine-as-imports = true
# Place "closer" imports (fewer . characters, most local) before "further" imports.
relative-imports-order = "closest-to-furthest"

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"

[tool.ruff.lint.pep8-naming]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
classmethod-decorators = ["classmethod", "pydantic.validator", "pydantic.root_validator"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.ruff.lint.flake8-annotations]
# Suppress ANN000-level violations for arguments matching the "dummy" variable regex (like _).
suppress-dummy-args = true

[tool.black]
line-length = 120
target-version = ["py311"]
# Colored output.
color = true
extend-exclude = '''
(
  .*_pb2_grpc.py # exclude autogenerated GRPC files anywhere in the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
  | .*_pb2.pyi  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''


[tool.pytest.ini_options]
pythonpath = [
    "src"
]

[tool.pyright]
typeCheckingMode = "standard"
include = ["src"]
exclude = [
    "src/experimental",
    "src/grpc_proto",
    "src/**/container.py",
]
reportMissingImports = "error"
reportMissingTypeStubs = false
pythonVersion = "3.11"